proofs verify now
